requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
gspread==6.0.2
google-auth==2.29.0
beautifulsoup4==4.12.3
requests==2.31.0
gspread==6.0.2
google-auth==2.29.0
